kafka_services/consumer/requirements.txt
kafka-python==2.0.2
sqlalchemy>=1.4.0
psycopg2-binary>=2.9.0
python-dotenv>=0.19.0
prometheus-client>=0.14.0
ratelimit>=2.2.1
backoff>=2.1.0